+++ R/test_diseasystore.R
@@ -159,6 +159,13 @@ test_diseasystore <- function(
 
 
 
+  # Make a little helper function to clean up the connection after each test
+  # and check that intermediate dbplyr and SCDB tables (e.g. from dplyr::compute() calls) are cleaned up
+  connection_clean_up <- function(conn) {
+    testthat::expect_identical(nrow(SCDB::get_tables(conn, show_temporary = TRUE, pattern = "dbplyr_")), 0L)
+    testthat::expect_identical(nrow(SCDB::get_tables(conn, show_temporary = TRUE, pattern = "SCDB_")), 0L)
+    DBI::dbDisconnect(conn)
+  }
 
   #     ######## ########  ######  ########  ######     ########  ########  ######   #### ##    ##  ######
   #        ##    ##       ##    ##    ##    ##    ##    ##     ## ##       ##    ##   ##  ###   ## ##    ##
@@ -197,6 +204,7 @@ test_diseasystore <- function(
     checkmate::expect_date(ds$min_start_date, upper = lubridate::today())
     checkmate::expect_date(ds$max_end_date,   upper = lubridate::today())
 
+    # Clean up
     rm(ds)
     invisible(gc())
   })
@@ -227,6 +235,7 @@ test_diseasystore <- function(
     feature <- testthat::expect_no_error(ds$available_features[[1]])
     testthat::expect_no_error(ds$get_feature(feature))
 
+    # Clean up
     rm(ds)
     invisible(gc())
   })
@@ -247,6 +256,7 @@ test_diseasystore <- function(
 
     testthat::expect_no_error(ds$get_feature(ds$available_features[[1]]))
 
+    # Clean up
     rm(ds)
     invisible(gc())
   })
@@ -306,7 +316,12 @@ test_diseasystore <- function(
           ),
           expr = {
             feature_checksums <- ds$get_feature(.x, start_date = test_start_date, end_date = test_end_date) |>
-              SCDB::digest_to_checksum() |>
+              SCDB::digest_to_checksum()
+
+            # digest_to_checksum() creates an intermediary table in SQLite
+            if (inherits(conn, "SQLiteConnection")) SCDB::defer_db_cleanup(feature_checksums)
+
+            feature_checksums <- feature_checksums |>
               dplyr::pull("checksum") |>
               sort()
 
@@ -366,12 +381,17 @@ test_diseasystore <- function(
         # Copy to remote and continue checks
         if (!inherits(reference, "tbl_sql") ||
               (inherits(reference, "tbl_sql") && !identical(dbplyr::remote_con(reference), conn))) {
-          reference <- dplyr::copy_to(conn, df = reference, name = SCDB::unique_table_name("ds"))
+          reference <- dplyr::copy_to(conn, df = reference, name = SCDB::unique_table_name("ds_reference"))
           SCDB::defer_db_cleanup(reference)
         }
 
         reference_checksums <- reference |>
-          SCDB::digest_to_checksum() |>
+          SCDB::digest_to_checksum()
+
+        # digest_to_checksum() creates an intermediary table in SQLite
+        if (inherits(conn, "SQLiteConnection")) SCDB::defer_db_cleanup(reference_checksums)
+
+        reference_checksums <- reference_checksums |>
           dplyr::pull("checksum") |>
           sort()
 
@@ -380,6 +400,8 @@ test_diseasystore <- function(
 
       })
 
+      # Clean up
+      connection_clean_up(conn)
       rm(ds)
       invisible(gc())
     }
@@ -410,7 +432,12 @@ test_diseasystore <- function(
           expr = {
 
             feature_checksums <- ds$get_feature(.x, start_date = test_start_date, end_date = test_end_date) |>
-              SCDB::digest_to_checksum() |>
+              SCDB::digest_to_checksum()
+
+            # digest_to_checksum() creates an intermediary table in SQLite
+            if (inherits(conn, "SQLiteConnection")) SCDB::defer_db_cleanup(feature_checksums)
+
+            feature_checksums <- feature_checksums |>
               dplyr::pull("checksum") |>
               sort()
 
@@ -430,12 +457,17 @@ test_diseasystore <- function(
         # Copy to remote and continue checks
         if (!inherits(reference, "tbl_sql") ||
               (inherits(reference, "tbl_sql") && !identical(dbplyr::remote_con(reference), conn))) {
-          reference <- dplyr::copy_to(conn, df = reference, name = SCDB::unique_table_name("ds"))
+          reference <- dplyr::copy_to(conn, df = reference, name = SCDB::unique_table_name("ds_reference"))
           SCDB::defer_db_cleanup(reference)
         }
 
         reference_checksums <- reference |>
-          SCDB::digest_to_checksum() |>
+          SCDB::digest_to_checksum()
+
+        # digest_to_checksum() creates an intermediary table in SQLite
+        if (inherits(conn, "SQLiteConnection")) SCDB::defer_db_cleanup(reference_checksums)
+
+        reference_checksums <- reference_checksums |>
           dplyr::pull("checksum") |>
           sort()
 
@@ -448,6 +480,8 @@ test_diseasystore <- function(
 
       })
 
+      # Clean up
+      connection_clean_up(conn)
       rm(ds)
       invisible(gc())
     }
@@ -510,6 +544,8 @@ test_diseasystore <- function(
 
         })
 
+      # Clean up
+      connection_clean_up(conn)
       rm(ds)
       invisible(gc())
     }
@@ -603,6 +639,7 @@ test_diseasystore <- function(
         })
 
       # Clean up
+      connection_clean_up(conn)
       rm(ds)
       invisible(gc())
     }